requirements.txt
GPUtil==1.4.0
psutil==5.8.0
pyserial==3.5
